Write out tests for backup deletion
- refactoring the "repo with history" fixture for shared use
- fixing some implementation mistakes with the fastforward stump
